ui fixes and caves refractoring

+++ World/Map/WorldMap.gd
@@ -27,7 +27,7 @@ enum Tiles {
 }
 
 func _input(event):
-	if not PlayerData.interactive_screen_mode and not PlayerData.viewInventoryMode and has_node("/root/World"):
+	if not PlayerData.interactive_screen_mode and not PlayerData.viewInventoryMode and not PlayerData.viewSaveAndExitMode and has_node("/root/World"):
 		if event.is_action_pressed("open_map"):
 			Server.player_node.actions.destroy_placable_object()
 			Server.world.get_node("WorldAmbience").hide()
